add stripe and change api/bookingController and web/bookingController

+++ app/Http/Controllers/Web/BokkingController.php
@@ -15,20 +15,17 @@ use Illuminate\Support\Facades\Response;
 
 class BokkingController extends Controller
 {
-
     public function index(Request $request)
     {
-        $user = $request->user(); // Obtén al usuario autenticado
+        $user = $request->user();
         $bookings = Bokking::where('user_id', $user->id)
-                           ->where('date', '>=', now()->toDateString()) // Solo reservas futuras
-                           ->orderBy('date', 'asc') // Ordenar por fecha de manera ascendente
-                           ->orderBy('time', 'asc') // Luego, ordenar por hora de manera ascendente
-                           ->get(); // Obtén las reservas del usuario actual
-    
+                           ->where('date', '>=', now()->toDateString())
+                           ->orderBy('date', 'asc')
+                           ->orderBy('time', 'asc')
+                           ->get();
+
         return view('dashboard', compact('bookings'));
     }
-    
-    
 
     public function create()
     {
@@ -38,17 +35,14 @@ class BokkingController extends Controller
 
     public function store(Request $request)
     {
-        // Validación de campos
         $request->validate([
             'date' => 'required|date|after_or_equal:today|before_or_equal:' . now()->addDays(30)->toDateString(),
             'time' => 'required|in:10:00,11:00,12:00,13:00',
             'horse_id' => 'required|exists:horses,id',
             'email' => 'required|email',
             'comments' => 'nullable|string',
-
         ]);
-    
-        // Verificar si las horas del día ya han pasado
+
         $selectedDate = new \DateTime($request->date);
         $currentTime = now();
         if ($selectedDate->format('Y-m-d') == $currentTime->format('Y-m-d')) {
@@ -59,20 +53,17 @@ class BokkingController extends Controller
                 return redirect()->back()->withErrors(['time' => 'Las horas de los turnos para hoy ya han terminado. Vuelve mañana.'])->withInput();
             }
         }
-    
-        // Verificar que la fecha seleccionada sea un sábado o domingo
+
         $date = new \DateTime($request->date);
         if (!in_array($date->format('N'), [6, 7])) {
             return redirect()->back()->withErrors(['date' => 'Las reservas solo están permitidas los sábados y domingos.'])->withInput();
         }
-    
-        // Verificar que el caballo no esté enfermo
+
         $horse = Horse::findOrFail($request->horse_id);
         if ($horse->sick) {
             return redirect()->back()->withErrors(['horse_id' => 'No se puede reservar un caballo que esté enfermo.'])->withInput();
         }
-    
-        // Verificar que el usuario no tenga una reserva en el mismo turno
+
         $existingBooking = Bokking::where('date', $request->date)
             ->where('time', $request->time)
             ->where('user_id', Auth::id())
@@ -80,8 +71,7 @@ class BokkingController extends Controller
         if ($existingBooking) {
             return redirect()->back()->withErrors(['time' => 'Ya tienes una reserva en este turno.'])->withInput();
         }
-    
-        // Verificar que el caballo no esté reservado en el mismo turno
+
         $existingBooking = Bokking::where('date', $request->date)
             ->where('time', $request->time)
             ->where('horse_id', $request->horse_id)
@@ -89,14 +79,13 @@ class BokkingController extends Controller
         if ($existingBooking) {
             return redirect()->back()->withErrors(['horse_id' => 'El caballo ya está reservado en este turno.'])->withInput();
         }
-    
-        // Verificar que el usuario no tenga otra reserva en las últimas 2 horas del mismo día
+
         $time = new \DateTime($request->time);
         $earlierTime = (clone $time)->modify('-2 hours')->format('H:i');
-    
+
         $bookingTimes = ['10:00', '11:00', '12:00', '13:00'];
         $currentIndex = array_search($request->time, $bookingTimes);
-    
+
         for ($i = max(0, $currentIndex - 2); $i <= $currentIndex; $i++) {
             if ($i != $currentIndex) {
                 $lastBooking = Bokking::where('user_id', Auth::id())
@@ -109,8 +98,7 @@ class BokkingController extends Controller
                 }
             }
         }
-    
-        // Pasar los datos de la reserva a la vista payment
+
         $bookingData = [
             'date' => $request->date,
             'time' => $request->time,
@@ -119,10 +107,11 @@ class BokkingController extends Controller
             'email' => $request->email,
             'horse_price' => $horse->price,
         ];
-    
+          
+       
+
         return view('payment', compact('bookingData'));
     }
-    
 
     public function show($id)
     {
@@ -133,10 +122,17 @@ class BokkingController extends Controller
     public function edit($id)
     {
         $booking = Bokking::where('user_id', Auth::id())->where('id', $id)->firstOrFail();
+        $currentDateTime = now();
+        $bookingDateTime = \Carbon\Carbon::parse($booking->date . ' ' . $booking->time);
+
+        if ($currentDateTime->greaterThanOrEqualTo($bookingDateTime)) {
+            return redirect()->route('bookings.index')->withErrors(['message' => 'No se puede editar una reserva pasada.']);
+        }
+
         $horses = Horse::all();
-        return view('Booking.edit', compact('booking','horses'));
+        return view('Booking.edit', compact('booking', 'horses'));
     }
-    
+
     public function update(Request $request, $id)
     {
         $request->validate([
@@ -147,8 +143,13 @@ class BokkingController extends Controller
         ]);
 
         $booking = Bokking::where('user_id', Auth::id())->where('id', $id)->firstOrFail();
+        $currentDateTime = now();
+        $bookingDateTime = \Carbon\Carbon::parse($booking->date . ' ' . $booking->time);
 
-       // Verificar si las horas del día ya han pasado
+        if ($currentDateTime->greaterThanOrEqualTo($bookingDateTime)) {
+            return redirect()->route('bookings.index')->withErrors(['message' => 'No se puede actualizar una reserva pasada.']);
+        }
+
         $selectedDate = new \DateTime($request->date);
         $currentTime = now();
         if ($selectedDate->format('Y-m-d') == $currentTime->format('Y-m-d')) {
@@ -160,19 +161,16 @@ class BokkingController extends Controller
             }
         }
 
-        // Verificar que la fecha seleccionada sea un sábado o domingo
         $date = new \DateTime($request->date);
         if (!in_array($date->format('N'), [6, 7])) {
             return redirect()->back()->withErrors(['date' => 'Las reservas solo están permitidas los sábados y domingos.'])->withInput();
         }
 
-        // Verificar que el caballo no esté enfermo
         $horse = Horse::findOrFail($request->horse_id);
         if ($horse->sick) {
             return redirect()->back()->withErrors(['horse_id' => 'No se puede reservar un caballo que esté enfermo.'])->withInput();
         }
 
-        // Verificar que no haya más de 5 reservas en el mismo turno
         $bookingsCount = Bokking::where('date', $request->date)
             ->where('time', $request->time)
             ->count();
@@ -180,7 +178,6 @@ class BokkingController extends Controller
             return redirect()->back()->withErrors(['time' => 'El turno ya está completo.'])->withInput();
         }
 
-        // Verificar que el caballo no esté reservado en el mismo turno
         $existingBooking = Bokking::where('date', $request->date)
             ->where('time', $request->time)
             ->where('horse_id', $request->horse_id)
@@ -189,9 +186,6 @@ class BokkingController extends Controller
             return redirect()->back()->withErrors(['horse_id' => 'El caballo ya está reservado en este turno.'])->withInput();
         }
 
-   
-        
-        // Verificar que el usuario no tenga otra reserva en las últimas 2 horas del mismo día
         $time = new \DateTime($request->time);
         $earlierTime = (clone $time)->modify('-2 hours')->format('H:i');
 
@@ -210,9 +204,7 @@ class BokkingController extends Controller
                 }
             }
         }
-        
 
-        // Actualizar la reserva
         $booking->fill($request->all());
         $booking->save();
 
@@ -222,6 +214,13 @@ class BokkingController extends Controller
     public function destroy($id)
     {
         $booking = Bokking::where('user_id', Auth::id())->where('id', $id)->firstOrFail();
+        $currentDateTime = now();
+        $bookingDateTime = \Carbon\Carbon::parse($booking->date . ' ' . $booking->time);
+
+        if ($currentDateTime->greaterThanOrEqualTo($bookingDateTime)) {
+            return redirect()->route('bookings.index')->withErrors(['message' => 'No se puede eliminar una reserva pasada.']);
+        }
+
         $booking->delete();
 
         return redirect()->route('bookings.index')->with('success', 'Reserva eliminada exitosamente.');
@@ -231,29 +230,24 @@ class BokkingController extends Controller
     {
         $booking = Bokking::findOrFail($id);
 
-        // Cargar la vista del detalle de la reserva para el PDF
         $html = view('Booking.show_pdf', compact('booking'))->render();
 
-        // Crear una instancia de Dompdf con opciones
         $options = new Options();
         $options->set('isHtml5ParserEnabled', true);
         $options->set('isPhpEnabled', true);
         $dompdf = new Dompdf($options);
 
-        // Cargar el contenido HTML en Dompdf
         $dompdf->loadHtml($html);
 
-        // Renderizar el PDF
         $dompdf->render();
 
-        // Obtener el contenido del PDF
         $pdfContent = $dompdf->output();
 
-        // Enviar la respuesta con el archivo PDF
         $response = Response::make($pdfContent);
         $response->header('Content-Type', 'application/pdf');
-        $response->header('Content-Disposition', 'inline; filename="booking_details_' . $id . '.pdf"');
+        $response->header('Content-Disposition', 'inline; filename="Detalles_reserva' . $id . '.pdf"');
 
         return $response;
     }
 }
+?>
